+++ src/Captura.Core/Captura.Core.csproj
@@ -102,6 +102,7 @@
     <Compile Include="Services\ServiceName.cs" />
     <Compile Include="Services\ServiceProvider.cs" />
     <Compile Include="Settings.cs" />
+    <Compile Include="TranslationSource.cs" />
     <Compile Include="ViewModels\AudioViewModel.cs" />
     <Compile Include="ViewModels\MainViewModel.cs" />
     <Compile Include="ViewModels\Properties.cs" />
