--- src/Captura/Languages/default.xaml
@@ -1,40 +1,0 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:v="clr-namespace:System;assembly=mscorlib">
-    <v:String x:Key="s_Title">Captura</v:String>
-    <v:String x:Key="s_About">About</v:String>
-    <v:String x:Key="s_Recent">Recent</v:String>
-    <v:String x:Key="s_Other">Other</v:String>
-    <v:String x:Key="s_Audio">Audio</v:String>
-    <v:String x:Key="s_Video">Video</v:String>
-    <v:String x:Key="s_Gif">Gif</v:String>
-    <v:String x:Key="s_GifUnconstrained">Unconstrained FrameRate</v:String>
-    <v:String x:Key="s_SaveTo">Save to</v:String>
-    <v:String x:Key="s_as">as</v:String>
-    <v:String x:Key="s_Resize">Resize</v:String>
-    <v:String x:Key="s_Repeat">Repeat</v:String>
-    <v:String x:Key="s_ScreenShot">ScreenShot</v:String>
-    <v:String x:Key="s_RecordStop">Record|Stop</v:String>
-    <v:String x:Key="s_PauseResume">Pause|Resume</v:String>
-    <v:String x:Key="s_Minimize">Minimize</v:String>
-    <v:String x:Key="s_Close">Close</v:String>
-    <v:String x:Key="s_Refresh">Refresh</v:String>
-    <v:String x:Key="s_OpenOutputFolder">Open Output Folder</v:String>
-    <v:String x:Key="s_Configure">Configure</v:String>
-    <v:String x:Key="s_IncludeCursor">Include Cursor</v:String>
-    <v:String x:Key="s_IncludeMouseClicks">Include Mouse Clicks</v:String>
-    <v:String x:Key="s_IncludeKeystrokes">Include Keystrokes</v:String>
-    <v:String x:Key="s_VideoSource">Video Source</v:String>
-    <v:String x:Key="s_VideoEncoder">Video Encoder</v:String>
-    <v:String x:Key="s_EncodeMp3">Encode Mp3 when only Audio</v:String>
-    <v:String x:Key="s_MicSource">Microphone Source</v:String>
-    <v:String x:Key="s_LoopbackSource">Wasapi Loopback Source</v:String>
-    <v:String x:Key="s_FrameRate">Frame Rate</v:String>
-    <v:String x:Key="s_Quality">Quality</v:String>
-    <v:String x:Key="s_BgColor">Background Color</v:String>
-    <v:String x:Key="s_StartDelay">Start Delay (in ms)</v:String>
-    <v:String x:Key="s_CaptureDuration">Capture Duration (in seconds)</v:String>
-    <v:String x:Key="s_MinCaptureStart">Minimize on Capture Start</v:String>
-    <v:String x:Key="s_MinTray">Minimize to System Tray</v:String>
-    <v:String x:Key="s_OutputFolder">Output Folder</v:String>
-</ResourceDictionary>
