--- src/Captura/Languages/lang.hi-IN.xaml
@@ -1,10 +1,0 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:v="clr-namespace:System;assembly=mscorlib">
-    <v:String x:Key="s_Other">अन्य</v:String>
-    <v:String x:Key="s_About">परिचय</v:String>
-    <v:String x:Key="s_Audio">ध्वनि</v:String>
-    <v:String x:Key="s_Video">वीडियो</v:String>
-    <v:String x:Key="s_ScreenShot">स्क्रीनशॉट</v:String>
-    <v:String x:Key="s_OutputFolder">आउटपुट फ़ोल्डर</v:String>
-</ResourceDictionary>
