--- src/Captura/Languages/lang.ml-IN.xaml
@@ -1,10 +1,0 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:v="clr-namespace:System;assembly=mscorlib">
-    <v:String x:Key="s_Other">മറ്റ്</v:String>
-    <v:String x:Key="s_Audio">ശബ്ദം</v:String>
-    <v:String x:Key="s_Video">വീഡിയോ</v:String>
-    <v:String x:Key="s_Recent">സമീപകാല</v:String>
-    <v:String x:Key="s_ScreenShot">സ്ക്രീൻഷോട്ട്</v:String>
-    <v:String x:Key="s_OutputFolder">ഔട്ട്പുട്ട് ഫോൾഡർ</v:String>
-</ResourceDictionary>
