+++ src/Captura/ViewModels/AboutViewModel.cs
@@ -1,58 +1,22 @@
 ﻿using Captura.ViewModels;
-using System;
+using System.Collections.ObjectModel;
 using System.Globalization;
-using System.Linq;
-using System.Threading;
-using System.Windows;
 
 namespace Captura
 {
     public class AboutViewModel : ViewModelBase
     {
-        public static CultureInfo DefaultLanguage { get; } = new CultureInfo("en-US");
-
-        public static CultureInfo[] Languages { get; } =
-        {
-            DefaultLanguage,
-            new CultureInfo("hi-IN"),
-            new CultureInfo("ml-IN")
-        };
+        public ObservableCollection<CultureInfo> Languages { get; }
 
         public AboutViewModel()
         {
-            // Initial Language Setup
-            Language = Language;
+            Languages = TranslationSource.Instance.AvailableCultures;            
         }
         
         public CultureInfo Language
         {
-            get { return new CultureInfo(Settings.Instance.Language); }
-            set
-            {
-                value = value ?? DefaultLanguage;
-                                
-                Thread.CurrentThread.CurrentUICulture = value;
-
-                Settings.Instance.Language = value.Name;
-
-                var dict = new ResourceDictionary()
-                {
-                    Source = new Uri($"Languages/lang.{value.Name}.xaml", UriKind.Relative)
-                };
-
-                var mergedDicts = Application.Current.Resources.MergedDictionaries;
-
-                var oldDict = mergedDicts.Where(d => d.Source != null && d.Source.OriginalString.StartsWith("Languages/lang."))
-                                         .First();
-
-                if (oldDict?.Source == dict.Source)
-                    return;
-
-                mergedDicts.Add(dict);
-
-                if (oldDict != null)
-                    mergedDicts.Remove(oldDict);
-            }
+            get => TranslationSource.Instance.CurrentCulture;
+            set => TranslationSource.Instance.CurrentCulture = value;
         }
     }
 }
